+++ WalletAdmin/Entidades/Tabela_Pessoas.cs
@@ -10,9 +10,9 @@ namespace WalletAdmin.Entidades
         public virtual int PES_CODIGO { get; set; }
         public virtual string PES_NOME { get; set; }
         public virtual string PES_EMAIL { get; set; }
-        public virtual double PES_SALARIO { get; set; }
-        public virtual double PES_LIMITE { get; set; }
-        public virtual double PES_MINIMO { get; set; }
-        public virtual double PES_SALDO { get; set; }
+        public virtual decimal PES_SALARIO { get; set; }
+        public virtual decimal PES_LIMITE { get; set; }
+        public virtual decimal PES_MINIMO { get; set; }
+        public virtual decimal PES_SALDO { get; set; }
     }
 }
